Moved database update of seats in ConfirmationActivity. SeatsActivity now displays a message if the room is full.
